Refactor to separate hardware-specific, send, and receive code.
